# Reproduce relationship performance problem