--- a/app/adapters/application.js
+++ b/app/adapters/application.js
@@ -1,0 +1,4 @@
+import DS from 'ember-data';
+
+export default DS.JSONAPIAdapter.extend({
+});