--- a/app/models/car.js
+++ b/app/models/car.js
@@ -1,0 +1,9 @@
+import DS from 'ember-data';
+
+const { Model, belongsTo, hasMany } = DS;
+
+export default Model.extend({
+  type: belongsTo('type', { async: false }),
+  size: belongsTo('size', { async: false }),
+  colors: hasMany('color', { async: false })
+});