--- a/app/serializers/application.js
+++ b/app/serializers/application.js
@@ -1,0 +1,4 @@
+import DS from 'ember-data';
+
+export default DS.JSONAPISerializer.extend({
+});