… package.json
… public/posts