## Ember Data 3.4.4

```
Pushing 2000 cars into the store: 113.052734375ms
Loop over 2000 cars and access all related data: 3738.7890625ms
```

## Ember Data 3.13.1

```
Pushing 2000 cars into the store: 318.47900390625ms
Loop over 2000 cars and access all related data: 122.94873046875ms
```

From this example, it seems like the time tripled to push the cars into the store, but accessing all of the data significantly decreases. These results seem consistent with relationships being lazy up until v3.4.4.